test: extract execution-model tests; add Reset and concurrency scenarios
